refactor: cut comments and split oversize files per PR review
Cleaned up comment density across PR-touched files (only kept comments
that explain a non-obvious *why*: a constraint, an invariant, a
workaround for a specific bug, or behavior that would surprise the
reader). Trimmed file headers, restated-section dividers, and obvious
type-field captions.

Files cleaned (line / comment-line counts after):
- cmd/buildtemplates/image.go (199 / 23)
- e2e/conversations_lifecycle_test.go (196 / 10)
- e2e/skills_passthrough_test.go (171 / 3)
- e2e/approvals_e2e_test.go (161 / 0)
- e2e/webhook_sequence_e2e_test.go (217 / 7)
- e2e/fakebridge/* (split, see below)
- internal/sandbox/bridge_codegen_smoke_test.go (197 / 4)
- internal/sandbox/pusher_agent_config_test.go (185 / 0)
- internal/sandbox/pusher_dead_features_removed_test.go (148 / 4)
- internal/sandbox/image_runtime_test.go (107 / 3)
- internal/sandbox/pusher_config.go (54 / 4)
- internal/sandbox/orchestrator_types.go (98 / 4)
- internal/sandbox/orchestrator_create_dedicated.go (173 / 2)
- internal/sandbox/daytona/driver_snapshot.go (186 / 8)
- internal/model/agent.go (Harness-field block trimmed)
- internal/model/org.go (subagent migration block simplified)

Files split (each Go file now under the 300-line CI ceiling without
adding to the allowlist):

- e2e/fakebridge/fakebridge.go (549) ->
    fakebridge.go (Server struct + lifecycle, 144),
    routes.go (HTTP handlers, 271),
    webhook.go (signed/unsigned PostWebhook, 81)

- e2e/conversations_lifecycle_test.go (401) ->
    conversations_lifecycle_test.go (Test* funcs, 196),
    conversations_lifecycle_helpers_test.go (newFakeBridgeHarness, 132)

- e2e/approvals_e2e_test.go (344) ->
    approvals_e2e_test.go (Test* funcs + runApprovalRoundtrip, 161),
    approvals_e2e_helpers_test.go (harness + waitForType/readAll, 172)

- e2e/webhook_sequence_e2e_test.go (361) ->
    webhook_sequence_e2e_test.go (Test* funcs, 217),
    webhook_sequence_e2e_helpers_test.go (harness builders, 107)

- internal/sandbox/pusher_agent_config_test.go (315) ->
    pusher_agent_config_test.go (field-pass-through table, 185),
    pusher_harness_test.go (pushHarnessFixture + HarnessFromAgent, 103)

scripts/check-go-file-length.sh now reports zero violations against the
unchanged allowlist. Tests pass: go test ./internal/bridge/...
./internal/sandbox/... ./internal/handler/... ./e2e/...
