+++ e2e/conversations_lifecycle_test.go
@@ -1,36 +1,7 @@
-// Wave 3 e2e: full conversation lifecycle against the fakebridge.
-//
-// These tests bring up the hiveloop conversation handler in-process, wire
-// it to a real Postgres + Redis (via the existing testHarness), and route
-// bridge calls to a scriptable fake bridge that speaks the new ACP-harness
-// wire contract. We can't spin a real bridge process — the actual binary
-// is not yet released against this contract.
-//
-// Required infra:
-//   DATABASE_URL  → Postgres reachable (defaults to localhost:5433/hiveloop_test)
-//   REDIS_ADDR    → Redis reachable (defaults to localhost:6379)
-// Tests skip with t.Skipf if either is unavailable, matching the existing
-// e2e harness pattern.
-//
-// Coverage of the deliverable spec:
-//   - Sandbox + AgentConversation rows are created in real Postgres.
-//   - Sandbox.BridgeURL points at the fakebridge so SendMessage/Stream/End
-//     hit it.
-//   - SendMessage drives the fakebridge to return scripted SSE events; the
-//     fakebridge then POSTs those events back as a webhook (matching the
-//     real bridge → hiveloop event-flow), and we assert the SSE consumer
-//     receives them in order with monotonic sequence_numbers.
-//   - End flips AgentConversation.status to "ended".
-//
-// We deliberately do NOT exercise the orchestrator's CreateDedicatedSandbox
-// flow here — that requires a real provider/turso. The sandbox-creation
-// codepath is covered by internal/sandbox tests.
 package e2e
 
 import (
-	"encoding/base64"
 	"encoding/json"
-	"fmt"
 	"io"
 	"net/http"
 	"net/http/httptest"
@@ -38,169 +9,22 @@ import (
 	"testing"
 	"time"
 
-	"github.com/go-chi/chi/v5"
-	"github.com/google/uuid"
-
 	"github.com/usehiveloop/hiveloop/e2e/fakebridge"
-	"github.com/usehiveloop/hiveloop/internal/crypto"
-	"github.com/usehiveloop/hiveloop/internal/handler"
-	"github.com/usehiveloop/hiveloop/internal/middleware"
 	"github.com/usehiveloop/hiveloop/internal/model"
-	"github.com/usehiveloop/hiveloop/internal/streaming"
 )
 
-// fbHarness wires the hiveloop conversation handler + bridge webhook
-// handler to the fakebridge so tests can drive the full pipeline.
-type fbHarness struct {
-	*testHarness
-	encKey   *crypto.SymmetricKey
-	bridge   *fakebridge.Server
-	org      model.Org
-	agent    model.Agent
-	sandbox  model.Sandbox
-	conv     model.AgentConversation
-	router   *chi.Mux
-	eventBus *streaming.EventBus
-	secret   string
-}
-
-// newFakeBridgeHarness creates an org/agent/sandbox/conv pre-wired so
-// that hiveloop's conversation handler will route bridge calls to the
-// fakebridge URL. Caller must seed `harness.bridge.ScriptedSSE` etc.
-// before exercising the routes.
-func newFakeBridgeHarness(t *testing.T) *fbHarness {
-	t.Helper()
-	h := newHarness(t)
-	suffix := uuid.New().String()[:8]
-
-	keyBytes := make([]byte, 32)
-	for i := range keyBytes {
-		keyBytes[i] = byte(i + 17)
-	}
-	encKey, err := crypto.NewSymmetricKey(base64.StdEncoding.EncodeToString(keyBytes))
-	if err != nil {
-		t.Fatalf("symmetric key: %v", err)
-	}
-
-	bridgeSecret := "fb-secret-" + suffix
-	encryptedKey, err := encKey.EncryptString(bridgeSecret)
-	if err != nil {
-		t.Fatalf("encrypt: %v", err)
-	}
-
-	fb := fakebridge.New(t)
-	fb.SignSecret = []byte(bridgeSecret)
-
-	org := model.Org{Name: "fb-org-" + suffix}
-	h.db.Create(&org)
-	t.Cleanup(func() { h.db.Where("id = ?", org.ID).Delete(&model.Org{}) })
-
-	cred := model.Credential{
-		OrgID: org.ID, BaseURL: "https://api.openai.com", AuthScheme: "bearer",
-		ProviderID: "openai", EncryptedKey: []byte("enc"), WrappedDEK: []byte("dek"),
-	}
-	h.db.Create(&cred)
-	t.Cleanup(func() { h.db.Where("id = ?", cred.ID).Delete(&model.Credential{}) })
-
-	agent := model.Agent{
-		OrgID: &org.ID, Name: "fb-agent-" + suffix,
-		CredentialID: &cred.ID, SystemPrompt: "test", Model: "gpt-4o",
-	}
-	h.db.Create(&agent)
-	t.Cleanup(func() { h.db.Where("id = ?", agent.ID).Delete(&model.Agent{}) })
-
-	// Sandbox with BridgeURL pointing at fakebridge — pre-set
-	// BridgeURLExpiresAt far in the future to short-circuit refresh.
-	expiresAt := time.Now().Add(24 * time.Hour)
-	sb := model.Sandbox{
-		OrgID:                 &org.ID,
-		AgentID:               &agent.ID,
-		ExternalID:            "fb-ext-" + suffix,
-		BridgeURL:             fb.URL,
-		BridgeURLExpiresAt:    &expiresAt,
-		EncryptedBridgeAPIKey: encryptedKey,
-		Status:                "running",
-	}
-	h.db.Create(&sb)
-	t.Cleanup(func() { h.db.Where("id = ?", sb.ID).Delete(&model.Sandbox{}) })
-
-	conv := model.AgentConversation{
-		OrgID: org.ID, AgentID: agent.ID, SandboxID: sb.ID,
-		BridgeConversationID: "fb-conv-" + suffix, Status: "active",
-	}
-	h.db.Create(&conv)
-	t.Cleanup(func() {
-		h.db.Where("conversation_id = ?", conv.ID).Delete(&model.ConversationEvent{})
-		h.db.Where("id = ?", conv.ID).Delete(&model.AgentConversation{})
-	})
-
-	// Build EventBus on top of the real Redis client.
-	eventBus := streaming.NewEventBus(h.redisClient)
-
-	// Wire the conversation handler. Orchestrator/Pusher are not strictly
-	// needed for SendMessage/End/Stream because we pre-created the
-	// sandbox, but the handler signature requires them. We construct a
-	// minimal orchestrator with nil provider/turso since we only call
-	// GetBridgeClient on the pre-wired sandbox — that path needs only
-	// the encKey and the sandbox row.
-	convHandler := handler.NewConversationHandler(h.db, nil, nil, eventBus)
-
-	// Webhook handler routes incoming events into Redis (eventBus) and
-	// Postgres. Use the same encKey so signature verification works.
-	webhookHandler := handler.NewBridgeWebhookHandler(h.db, encKey, eventBus, nil)
-
-	r := chi.NewRouter()
-	r.Use(func(next http.Handler) http.Handler {
-		return http.HandlerFunc(func(w http.ResponseWriter, req *http.Request) {
-			req = middleware.WithOrg(req, &org)
-			next.ServeHTTP(w, req)
-		})
-	})
-	r.Route("/v1/conversations/{convID}", func(r chi.Router) {
-		r.Get("/stream", convHandler.Stream)
-	})
-	r.Post("/internal/webhooks/bridge/{sandboxID}", webhookHandler.Handle)
-
-	// Point the fakebridge at the webhook route on this router.
-	whSrv := httptest.NewServer(r)
-	t.Cleanup(whSrv.Close)
-	fb.WebhookURL = whSrv.URL + "/internal/webhooks/bridge/" + sb.ID.String()
-
-	return &fbHarness{
-		testHarness: h,
-		encKey:      encKey,
-		bridge:      fb,
-		org:         org,
-		agent:       agent,
-		sandbox:     sb,
-		conv:        conv,
-		router:      r,
-		eventBus:    eventBus,
-		secret:      bridgeSecret,
-	}
-}
-
-// TestConversationLifecycle_PushSendStreamEnd exercises:
-//   - fakebridge accepts a webhook batch carrying a scripted turn,
-//   - hiveloop persists non-chunk events to Postgres,
-//   - hiveloop publishes everything to Redis,
-//   - the SSE consumer drains them in order with monotonic seq numbers,
-//   - DELETE conv flips status="ended".
-//
-// We don't drive POST /v1/agents/{id}/conversations because that path
-// requires a real provider/turso to instantiate a sandbox. The sandbox
-// is pre-created in the harness with BridgeURL = fakebridge.URL.
+// TestConversationLifecycle_PushSendStreamEnd drives the full
+// webhook -> Redis -> SSE pipeline against the fakebridge: webhook batch
+// arrives, hiveloop publishes events to Redis, the SSE consumer drains
+// them with monotonic sequence numbers, and the ConversationEnded webhook
+// flips status="ended".
 func TestConversationLifecycle_PushSendStreamEnd(t *testing.T) {
 	fbh := newFakeBridgeHarness(t)
 
-	// Run the router on a real httptest server so SSE works correctly
-	// (httptest.ResponseRecorder doesn't support Flusher and races on
-	// concurrent reads/writes; a real server uses a proper net.Conn).
+	// Run on a real httptest server so SSE Flusher works (recorder doesn't).
 	srv := httptest.NewServer(fbh.router)
 	t.Cleanup(srv.Close)
 
-	// Open the SSE subscription as a real HTTP request, ask only for live
-	// events ("$" cursor) — that's the default but we set it explicitly.
 	sseReq, _ := http.NewRequest(http.MethodGet,
 		srv.URL+"/v1/conversations/"+fbh.conv.ID.String()+"/stream", nil)
 	sseReq.Header.Set("Accept", "text/event-stream")
@@ -211,8 +35,6 @@ func TestConversationLifecycle_PushSendStreamEnd(t *testing.T) {
 	}
 	defer sseResp.Body.Close()
 
-	// Channel-collected SSE event types. We read in a goroutine so the
-	// main thread can post the webhook and then read what arrived.
 	gotTypes := make(chan string, 16)
 	go func() {
 		defer close(gotTypes)
@@ -222,7 +44,6 @@ func TestConversationLifecycle_PushSendStreamEnd(t *testing.T) {
 			n, err := sseResp.Body.Read(buf)
 			if n > 0 {
 				acc += string(buf[:n])
-				// Split on blank-line boundaries; emit event type when seen.
 				for {
 					idx := strings.Index(acc, "\n\n")
 					if idx == -1 {
@@ -244,9 +65,8 @@ func TestConversationLifecycle_PushSendStreamEnd(t *testing.T) {
 		}
 	}()
 
-	// Give the subscriber time to attach to the Redis tap with cursor "$"
-	// before we publish. Events posted before the XREAD BLOCK starts may
-	// otherwise be missed.
+	// Wait for the Redis XREAD BLOCK to attach with cursor "$" before
+	// publishing — events posted before XREAD starts get missed.
 	time.Sleep(200 * time.Millisecond)
 
 	now := time.Now()
@@ -282,7 +102,6 @@ func TestConversationLifecycle_PushSendStreamEnd(t *testing.T) {
 		t.Fatalf("webhook POST: status=%d body=%s", status, body)
 	}
 
-	// Drain SSE event types as they arrive, with a per-test deadline.
 	expectedTypes := []string{"message_received", "response_chunk", "response_chunk", "response_chunk", "turn_completed"}
 	collected := make([]string, 0, len(expectedTypes))
 	deadline := time.After(5 * time.Second)
@@ -310,19 +129,8 @@ collectLoop:
 		}
 	}
 
-	// Note: with an EventBus wired, the webhook handler publishes only to
-	// Redis — Postgres persistence is the responsibility of the flusher
-	// goroutine (internal/streaming/flusher.go), which we don't run here.
-	// The webhook test (TestWebhook_PersistsEvents) covers the
-	// no-eventBus → direct-Postgres path. webhook_sequence_e2e_test.go
-	// covers the Redis Stream path including back-reading from the
-	// stream.
-
-	// End the conversation by simulating the bridge ConversationEnded
-	// webhook. This is the path the real bridge takes to terminate a
-	// conversation, and the webhook handler flips status="ended" on it.
-	// (Driving DELETE on hiveloop's End() handler would require a real
-	// orchestrator; out of scope for this lifecycle assertion.)
+	// End via ConversationEnded webhook — driving DELETE through hiveloop's
+	// End() handler would require a real orchestrator.
 	endEvents := []fakebridge.BridgeEvent{
 		{
 			EventID: "ev-end", EventType: "ConversationEnded",
@@ -341,20 +149,9 @@ collectLoop:
 	}
 }
 
-// TestConversationLifecycle_UpsertAgentNewWireShape proves the pusher's
-// UpsertAgent serialization (already covered by sandbox unit tests) lines
-// up with what the fakebridge sees when run in an e2e configuration. We
-// build a full agent definition and PUT it via the real bridge client to
-// the fakebridge, then assert the captured shape has no dead fields.
-//
-// This is a defensive belt-and-suspenders test that locks the wire shape
-// at the fakebridge boundary so future tests in this package can rely on
-// it without re-asserting the same invariants.
 func TestConversationLifecycle_UpsertAgentNewWireShape(t *testing.T) {
 	fb := fakebridge.New(t)
 
-	// Drive UpsertAgent directly — minimal AgentDefinition with all
-	// required fields filled.
 	body := `{
 		"id": "agent-1",
 		"name": "test",
@@ -387,7 +184,6 @@ func TestConversationLifecycle_UpsertAgentNewWireShape(t *testing.T) {
 		t.Errorf("model: got %q", def.Provider.Model)
 	}
 
-	// Decode the raw body and verify dead fields are not present.
 	var raw map[string]any
 	if err := json.Unmarshal(cap.UpsertAgentsRaw[0], &raw); err != nil {
 		t.Fatalf("decode raw: %v", err)
@@ -397,5 +193,4 @@ func TestConversationLifecycle_UpsertAgentNewWireShape(t *testing.T) {
 			t.Errorf("forbidden field %q present in upsert body", dead)
 		}
 	}
-	_ = fmt.Sprintf // keep import alive in helper file
 }
