+++ internal/sandbox/orchestrator_create_dedicated.go
@@ -107,9 +107,8 @@ func (o *Orchestrator) createSandbox(ctx context.Context, org *model.Org, agent 
 	sb.Status = "running"
 	sb.LastActiveAt = &now
 
-	// Ensure the per-harness config dirs exist on the running sandbox even when
-	// the snapshot was rebuilt without them (e.g. older snapshots used
-	// /home/daytona/.bridge). The new ACP-harness contract uses /work as HOME.
+	// Older snapshots may not have the harness config dirs; create them so
+	// the harnesses don't crash on first read.
 	if _, execErr := o.provider.ExecuteCommand(ctx, info.ExternalID, "mkdir -p /work/.claude /work/.opencode"); execErr != nil {
 		logging.Capture(ctx, fmt.Errorf("create bridge config dirs sandbox %s: %w", sb.ID, execErr))
 	}
