+++ internal/sandbox/orchestrator_types.go
@@ -14,9 +14,8 @@ import (
 )
 
 const (
-	// BridgePort is the listening port of the new ACP-harness bridge runtime.
-	// Kept at 25434 (the original hiveloop bridge port) — BRIDGE_LISTEN_ADDR
-	// instructs the new bridge binary to listen here instead of its 8080 default.
+	// BridgePort overrides the bridge binary's 8080 default via
+	// BRIDGE_LISTEN_ADDR. Kept at the original hiveloop port (25434).
 	BridgePort = 25434
 
 	bridgeHealthTimeout    = 90 * time.Second
@@ -34,9 +33,8 @@ func baseEnvVars(cfg *config.Config, bridgeAPIKey string, sandboxID uuid.UUID, w
 		"BRIDGE_LOG_FORMAT":            "json",
 		"BRIDGE_WEB_URL":               fmt.Sprintf("https://%s/spider", cfg.BridgeHost),
 		"HIVELOOP_SANDBOX_ID":          sandboxID.String(),
-		// New ACP-harness runtime contract: HOME=/work so bridge.db lives at
-		// /work/bridge.db and survives provider stop/start cycles. The Claude
-		// and OpenCode harnesses look for their config under these dirs.
+		// HOME=/work so bridge.db survives provider stop/start; the harnesses
+		// read their config from CLAUDE_CONFIG_DIR / OPENCODE_CONFIG_DIR.
 		"HOME":                "/work",
 		"CLAUDE_CONFIG_DIR":   "/work/.claude",
 		"OPENCODE_CONFIG_DIR": "/work/.opencode",
